src/requirements.txt
TkEasyGUI
Pillow

